land_lawma/requirements.txt
streamlit
pdfplumber
pytesseract
sentence-transformers
faiss-cpu
transformers
accelerate
bitsandbytes
torch
langchain
python-dotenv
tqdm
